remove server_common dependencies on blockserver and dbserver
