+++ BlockServer/core/config_list_manager.py
@@ -24,7 +24,7 @@ from BlockServer.core.config_list_manager_exceptions import InvalidDeleteExcepti
 from BlockServer.core.constants import DEFAULT_COMPONENT
 from BlockServer.core.file_path_manager import FILEPATH_MANAGER
 from BlockServer.core.inactive_config_holder import InactiveConfigHolder
-from BlockServer.core.macros import MACROS
+from server_common.helpers import MACROS
 from server_common.channel_access import ChannelAccess, verify_manager_mode
 from server_common.common_exceptions import MaxAttemptsExceededException
 from server_common.pv_names import BlockserverPVNames
