+++ block_server.py
@@ -41,7 +41,8 @@ from BlockServer.core.config_list_manager import ConfigListManager
 from BlockServer.core.file_path_manager import FILEPATH_MANAGER
 from BlockServer.core.inactive_config_holder import InactiveConfigHolder
 from BlockServer.core.ioc_control import IocControl
-from BlockServer.core.macros import BLOCK_PREFIX, CONTROL_SYSTEM_PREFIX, MACROS, PVPREFIX_MACRO
+from BlockServer.core.macros import BLOCK_PREFIX, PVPREFIX_MACRO
+from server_common.helpers import MACROS, CONTROL_SYSTEM_PREFIX
 from BlockServer.devices.devices_manager import DevicesManager
 from BlockServer.epics.archiver_manager import ArchiverManager
 from BlockServer.epics.gateway import Gateway
