+++ server_common/helpers.py
@@ -45,3 +45,19 @@ def get_macro_values():
 
 
 motor_in_set_mode = g.adv.motor_in_set_mode
+
+
+def _get_env_var(name):
+    try:
+        return os.environ[name]
+    except:
+        return ""
+
+
+MACROS = {
+    "$(MYPVPREFIX)": _get_env_var("MYPVPREFIX"),
+    "$(EPICS_KIT_ROOT)": _get_env_var("EPICS_KIT_ROOT"),
+    "$(ICPCONFIGROOT)": _get_env_var("ICPCONFIGROOT"),
+    "$(ICPVARDIR)": _get_env_var("ICPVARDIR"),
+}
+CONTROL_SYSTEM_PREFIX = MACROS["$(MYPVPREFIX)"] + "CS:"
